requirements.txt
setuptools>=70
flask==2.3.3
langchain==0.0.350
langchain-community==0.0.8
sentence-transformers==2.2.2
huggingface-hub==0.19.4
numpy==1.26.4
scikit-learn==1.4.2 
python-dotenv==1.0.0
google-generativeai==0.4.1
langchain-google-genai==1.0.1
pandas==2.2.2
pydantic==2.8.1
nltk
beautifulsoup4==4.12.2
chromadb==1.0.13
posthog==3.0.1
